+++ kwcoco/coco_dataset.py
@@ -50,13 +50,14 @@ An informal spec is as follows:
                 'channels': <ChannelSpec>,   # a string encoding
                 'width':     <int>    # pixel width of auxiliary image
                 'height':    <int>    # pixel height of auxiliary image
-                'base_to_aux': <TransformSpec>,  # tranform from "base" image space to auxiliary image space. (identity if unspecified)
+                'warp_aux_to_img': <TransformSpec>,  # tranform from "base" image space to auxiliary image space. (identity if unspecified)
             }, ...
         ]
 
         'video_id': str  # if this image is a frame in a video sequence, this id is shared by all frames in that sequence.
         'timestamp': str | int  # a iso-string timestamp or an integer in flicks.
         'frame_index': int  # ordinal frame index which can be used if timestamp is unknown.
+        'warp_img_to_vid': <TransformSpec>  # a transform image space to video space (identity if unspecified), can be used for sensor alignment or video stabilization
     }
 
     TransformSpec:
@@ -1006,15 +1007,17 @@ class MixinCocoExtras(object):
             cache_dpath = join(bundle_dpath, '_cache')
             fpath = join(bundle_dpath, 'data.kwcoco.json')
 
-            stamp = ub.CacheStamp('vidshape_stamp_v14', dpath=cache_dpath,
-                                  cfgstr=cfgstr, enabled=use_cache,
-                                  product=[fpath], verbose=100,
-                                  # meta=vidkw  # requires ubelt>=0.9.3
-                                  )
+            stamp = ub.CacheStamp(
+                'vidshape_stamp_v{:03d}'.format(toydata.TOYDATA_VERSION),
+                dpath=cache_dpath,
+                cfgstr=cfgstr, enabled=use_cache,
+                product=[fpath], verbose=100,
+                meta=vidkw
+            )
             print('stamp = {!r}'.format(stamp))
             if stamp.expired():
                 vidkw['dpath'] = bundle_dpath
-                vidkw.pop('bundle_dpath')
+                vidkw.pop('bundle_dpath', None)
                 self = toydata.random_video_dset(**vidkw)
                 print('self.fpath = {!r}'.format(self.fpath))
                 print('self.bundle_dpath = {!r}'.format(self.bundle_dpath))
@@ -2127,9 +2130,19 @@ class MixinCocoStats(object):
                 if 'area' not in ann:
                     # Use segmentation if available
                     if 'segmentation' in ann:
-                        import kwimage
-                        poly = kwimage.MultiPolygon.from_coco(ann['segmentation'])
-                        ann['area'] = float(poly.to_shapely().area)
+                        try:
+                            import kwimage
+                            poly = kwimage.MultiPolygon.from_coco(ann['segmentation'])
+                            ann['area'] = float(poly.to_shapely().area)
+                        except Exception:
+                            import warnings
+                            warnings.warn(ub.paragraph(
+                                '''
+                                Unable to coerce segmentation to a polygon.
+                                This may be indicative of a bug in
+                                `kwimage.MultiPolygon.coerce` or a misformatted
+                                segmentation
+                                '''))
                     else:
                         x, y, w, h = ann['bbox']
                         ann['area'] = w * h
@@ -2415,7 +2428,8 @@ class MixinCocoStats(object):
             statskw (dict): kwargs for :func:`kwarray.stats_dict`
 
         Returns:
-            Dict[str, Dict[str, Dict | ndarray]
+            Dict[str, Dict[str, Dict | ndarray]:
+                Stats are returned in width-height format.
 
         Example:
             >>> import kwcoco